BACKPORT: drm/i915: Broaden application of set-domain(GTT)
Previously, this was restricted to only operate on bound objects - to
make pointer access through the GTT to the object coherent with writes
to and from the GPU. A second usecase is drm_intel_bo_wait_rendering()
which at present does not function unless the object also happens to
be bound into the GGTT (on current systems that is becoming increasingly
rare, especially for the typical requests from mesa). A third usecase is
a future patch wishing to extend the coverage of the GTT domain to
include objects not bound into the GGTT but still in its coherent cache
domain. For the latter pair of requests, we need to operate on the
object regardless of its bind state.

v2: After discussion with Akash, we came to the conclusion that the
get-pages was required in order for accurate domain tracking in the
corner cases (like the shrinker) and also useful for ensuring memory
coherency with earlier cached CPU mmaps in case userspace uses exotic
cache bypass (non-temporal) instructions.

v3: Fix the inactive object check.

v4: Rebase to latest drm-intel-nightly codebase

BUG=chromium:602675
TEST=build chromeos-kernel-3_14 on samus, open Chrome and browser around.
Checked also that "Failed DMA_BUF_SYNC_END" error in not anymore in chrome log
file. Besides I ran (upstream) igt `prime_mmap_coherency --run-subtest
ioctl-errors`, which was failing before this patch.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Akash Goel <akash.goel@intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
(cherry picked from commit 43566dedde54f9729113f5f9fde77d53e75e61e9)
Signed-off-by: Marc Herbert <marc.herbert@intel.com>

Change-Id: I2bd115d87d6ba3bad0fbe352e91a647aaf4aafdc
Reviewed-on: https://chromium-review.googlesource.com/339030
Commit-Ready: Tiago Vignatti <tiago.vignatti@intel.com>
Tested-by: Tiago Vignatti <tiago.vignatti@intel.com>
Reviewed-by: Tiago Vignatti <tiago.vignatti@intel.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>