CHROMIUM: mali: register suspend notifier to runtime resume
This adds a pm notifier for mali to runtime resume the gpu device on
suspend prepare. Genpd will just leave the gpu device runtime
suspended on system suspend, but that causes issues for mali. This
works around that by making sure the mali gpu device is runtime
resumed on system suspend before the genpd prepare callback is called.

BUG=chrome-os-partner:51949
TEST=power_Resume

Change-Id: I4640b25b73da8feb812220cb085c7049b704e34c
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/339704
Reviewed-by: Douglas Anderson <dianders@chromium.org>