FROMLIST: loop: properly observe rotational flag of underlying device
The loop driver always declares the rotational flag of its device as
rotational, even when the device of the mapped file is nonrotational,
as is the case with SSDs or on tmpfs. This can confuse filesystem tools
which are SSD-aware; in my case I frequently forget to tell mkfs.btrfs
that my loop device on tmpfs is nonrotational, and that I really don't
need any automatic metadata redundancy.

The attached patch fixes this by introspecting the rotational flag of the
mapped file's underlying block device, if it exists. If the mapped file's
filesystem has no associated block device - as is the case on e.g. tmpfs -
we assume nonrotational storage. If there is a better way to identify such
non-devices I'd love to hear them.

[From: http://www.gossamer-threads.com/lists/linux/kernel/2301281]

BUG=b:25115863,b:25598659
TEST=Check loop device rotation sysfs attribue is now 0.

Change-Id: Id93da46be2b3c5945450240b163e3bb15adec6d0
Signed-off-by: Holger Hoffsttte <holger.hoffstaette@gmail.com>
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/347065