CHROMIUM: ARM: errata: set a chicken bit to fix A12 erratum
This adds CONFIG_ERRATA_FOOBAR (because this erratum doesn't have an
official ARM errata number) The erratum causes a hard CPU hang under
certain workloads, and setting this chicken bit appears to fix it with
no measureable loss in performance, so let's do it.  Similar to other
erratum workarounds we must set this bit upon resume as well.

Once we have an official erratum number we can send this upstream.

BUG=chrome-os-partner:52476
BUG=b:27112475
TEST=with config enabled, minnie boots and doesn't hang
TEST=with config enabled, minnie octane doesn't regress

Change-Id: I8de9ce425d52379416cdd6a82a3521b1c764b70d
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/342753
Reviewed-by: Dylan Reid <dgreid@chromium.org>