CHROMIUM: drm/rockchip: enable plane scaling
The VOP can handle scaling up or down by up to a factor of 8. Sadly, it
does not seem to handle vertical scale down factors near 2, 4, or 8.
This CL includes a workaround to slightly reduce the src height to
ensure the degenerate case is avoided.

TEST=plane_test
BUG=chromium:607241

Change-Id: I9d116a69bcb0d202a72a6ea1b068a767865ad108
Signed-off-by: Zach Reizner <zachr@google.com>
Reviewed-on: https://chromium-review.googlesource.com/343200
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>