CHROMIUM: mali: Fix driver behavior when runtime PM is disabled
Even if runtime PM is disabled, the driver should call
pm_runtime_get_sync() to keep the reference on power status, so that
when runtime PM is enabled asynchronously by some other entity, the
status is kept as necessary.

Moreover, do not disable the regulator when runtime PM is disabled, as
the runtime PM API specifies that the device should be powered on all
the time if runtime PM is disabled. To retain the ability to turn off
the regulator for system-wide suspend, implement system-wide suspend
callback which releases runtime PM count and disables clock if the GPU
was left powered on and then disables the regulator and resume callback
which only enables the regulator.

BUG=chrome-os-partner:51949
TEST=10 loops described in #72 with environment from #78

Change-Id: Iadd01a2fe855a5b0bc0d2092286ddb082af631fe
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/344821