CHROMIUM: ARM: dts: rockchip: configure properly Tiger ethernet PHY
The vcc_lan power rail going to the Ethernet PHY is connected to the
RK808 SW2 on Tiger/Fievel.
Put the right regulator configuration in the device-tree, so the
Ethernet PHY is powered.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=chrome-os-partner:53191
TEST=on a Tiger board, connect the wired ethernet to the on-board port,
do not see the stmmac driver complaining about the broken PHY and browse
the web.

Change-Id: Ia87d55ea3e56c5fca7dd735bd23322b1a37d93a9
Reviewed-on: https://chromium-review.googlesource.com/346494
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>