FROMLIST: net: stmmac: dwmac-rk: keep the PHY up for WoL
When suspending the machine, do not shutdown the external PHY by cutting
its regulator in the mac platform driver suspend code if Wake-on-Lan is enabled,
else it cannot wake us up.
In order to do this, split the suspend/resume callbacks from the
init/exit callbacks, so we can condition the power-down on the lack of
need to wake-up from the LAN but do it unconditionally when unloading the
module.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry-pick from commit 229666c14c75ae6c5016c49d8806438e843c6807 on
 davem/net-next tree)

BUG=chrome-os-partner:53191
TEST=manual: with series, on Tiger, suspend the machine and wake it up
from keyboard and using a WoL magic packet, and check the on-board
networking is still working.

Change-Id: Iecd6de1c5280b0c7002915592e81a4ca345fd997
Reviewed-on: https://chromium-review.googlesource.com/348853
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>