BACKPORT: dma-buf/fence: make fence context 64 bit v2
Fence contexts are created on the fly (for example) by the GPU scheduler used
in the amdgpu driver as a result of an userspace request. Because of this
userspace could in theory force a wrap around of the 32bit context number
if it doesn't behave well.

Avoid this by increasing the context number to 64bits. This way even when
userspace manages to allocate a billion contexts per second it takes more
than 500 years for the context number to wrap around.

v2: fix printf formats as well.

Signed-off-by: Christian König <christian.koenig@amd.com>
Reviewed-by: Gustavo Padovan <gustavo.padovan@collabora.co.uk>
Acked-by: Sumit Semwal <sumit.semwal@linaro.org>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Link: http://patchwork.freedesktop.org/patch/msgid/1464786612-5010-2-git-send-email-deathsimple@vodafone.de
(cherry picked from commit 76bf0db5543976ef50362db7071da367cb118532)
from topic/drm-misc-2016-06-07
Signed-off-by: Tarun <tarun.vyas@intel.com>

BUG=b:36446508
TEST=No regressions in KBL and APL based devices

 Conflicts: [Tarun: Retain the HEAD. Avoid changing anything]
	drivers/gpu/drm/amd/amdgpu/amdgpu.h
	drivers/gpu/drm/amd/amdgpu/amdgpu_sa.c

 Deleted by us: [Avoid adding this]
	drivers/gpu/drm/etnaviv/etnaviv_gpu.h
	[This one was renamed to sync_debug.h by "1fe82e2e1486
	staging/android: rename sync.h to sync_debug.h"]
	drivers/staging/android/sync.h

 Modified: [Merge commit "731c7d3a205ba89b4"]
 	drivers/staging/android/sync_debug.h

Change-Id: Ic7f8a575a45297c06561a059a7064468ed9480da