CHROMIUM: HACK: net: stmmac: reset PHY after WoL
Reset the PHY at resume when it is in low-power mode for Wake-on-Lan.

We might be able to do something more subtle in this case but so far, I
haven't been able to get it back in a fully working state after resuming
from low power mode unless I fully reset it.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=chrome-os-partner:53191
TEST=manual: with series, on Tiger, suspend the machine and wake it up
from keyboard and using a WoL magic packet, and check the on-board
networking is still working.

Change-Id: If7ccb61c562c3473516652bbb8e1496ed5a0c317
Reviewed-on: https://chromium-review.googlesource.com/348855
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>