FROMLIST: net: stmmac: allow to split suspend/resume from init/exit callbacks
Let the stmmac platform drivers provide dedicated suspend and resume
callbacks rather than always re-using the init and exits callbacks.
If the driver does not provide the suspend or resume callback, we fall
back to the old behavior trying to use exit or init.

This allows a specific platform to perform only a partial power-down on
suspend if Wake-on-Lan is enabled but always perform the full shutdown
sequence if the module is unloaded.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry-pick from commit cecbc5563a02289164fa6379130243cbe08b2dd6 on
 davem/net-next tree)

BUG=chrome-os-partner:53191
TEST=manual: with series, on Tiger, suspend the machine and wake it up
from keyboard and using a WoL magic packet, and check the on-board
networking is still working.

Change-Id: I2f2c5d4fb9b5852619858cdc45c3215ceecbfb29
Reviewed-on: https://chromium-review.googlesource.com/353341
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>