FROMLIST: ARM: dts: rockchip: add interrupt for Wake-on-Lan on RK3288
In order to use Wake-on-Lan on RK3288 integrated MAC, we need to wake-up
the CPU on the PMT interrupt when the MAC and the PHY are in low power mode.
Adding the interrupt declaration.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry-pick from commit d5bfbeb80973010dc87c82dc56f66e99c3e8c4da on
 davem/net-next tree)

BUG=chrome-os-partner:53191
TEST=manual: with series, on Tiger, suspend the machine and wake it up
from keyboard and using a WoL magic packet, and check the on-board networking
is still working.

Change-Id: I73f3f718fe759649381697ad26fd4ae2f4384225
Reviewed-on: https://chromium-review.googlesource.com/353342
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>