BACKPORT: kvm: vmx: hook preemption timer support
Hook the VMX preemption timer to the "hv timer" functionality added
by the previous patch.  This includes: checking if the feature is
supported, if the feature is broken on the CPU, the hooks to
setup/clean the VMX preemption timer, arming the timer on vmentry
and handling the vmexit.

A module parameter states if the VMX preemption timer should be
utilized.

Signed-off-by: Yunhong Jiang <yunhong.jiang@intel.com>
[Move hv_deadline_tsc to struct vcpu_vmx, use -1 as the "unset" value.
 Put all VMX bits here.  Enable it by default #yolo. - Paolo]
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>

BUG=b:120040011
TEST=Built

Conflicts:
   arch/x86/kvm/vmx.c

[delco: Conflict from absent neighboring content]

(cherry picked from commit 64672c95ea4c2f7096e519e826076867e8ef0938)
Signed-off-by: Matt Delco <delco@chromium.org>
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Change-Id: I46469b311df6a35978e03d1a5840739eb62af215
Reviewed-on: https://chromium-review.googlesource.com/1762326
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>