UPSTREAM: [media] cec-funcs.h: static inlines to pack/unpack CEC messages
This public header contains static inlines to pack and unpack CEC
messages. It is for use in both the kernel and in userspace.

Since the CEC framework will initially be in staging this header is
not yet in include/uapi. Once the framework is moved out of staging
this header should be moved to uapi at the same time.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>
(cherry picked from commit 50f7d5a65e5a31e92a9d46663770cd9509f939a2)
Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>

Change-Id: I20802172e62739c995ee256590d4aa7a3bb54d3f
Reviewed-on: https://chromium-review.googlesource.com/1021242
Commit-Ready: Sean Paul <seanpaul@google.com>
Tested-by: Sean Paul <seanpaul@google.com>
Reviewed-by: Sean Paul <seanpaul@google.com>