UPSTREAM: drm/i915: Initialize the PPS HW before its first use
The initial DPCD read for eDP detection involves using the PPS, but so
far we only initialized the PPS registers after the DPCD read. The
reason this was done so far is to preserve a possible LVDS PPS HW setup
if LVDS is detected but eDP is not. This is not an issue any more after
the previous patch, so we can move the init earlier now.

This was caught by CI with the PPS sanity checks in place and the
initial eDP DPCD readout waiting for the panel power cycle timeout
without the PPS registers being initialized.

CC: Ville Syrjl <ville.syrjala@linux.intel.com>
CC: Chris Wilson <chris@chris-wilson.co.uk>
Signed-off-by: Imre Deak <imre.deak@intel.com>
Reviewed-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Ville Syrjl <ville.syrjala@linux.intel.com>
Link: http://patchwork.freedesktop.org/patch/msgid/1466499109-20240-3-git-send-email-imre.deak@intel.com
(cherry picked from commit 0080b5da392243af673e752b5681e6388ede7da0)
Signed-off-by: Radhakrishna Sripada <radhakrishna.sripada@intel.com>

BUG=chrome-os-partner:53839
TEST=External monitors connected via MST hub or daisy-chain over DP should come up.

Change-Id: I8edebe85ec56b0fb810a768a75766ee6531cdbcb
Reviewed-on: https://chromium-review.googlesource.com/421996
Commit-Ready: Brian J Lovin <brian.j.lovin@intel.com>
Tested-by: Chenglei Ren <chenglei.ren@intel.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>