CHROMIUM: alt-syscall: Allow Android to set RT scheduler.
When running in the android container, Android needs to set RT
scheduler for some of its services.  Setting RT scheduler affects the
host so process inside container needs to have privilege targeted the
host to achieve that.  However this is not configurable nor supported
by kernel, so we use alt-syscall to provide Android the needed
functionality.

BUG=b:29250079
TEST=check thread prio while running audio from Android on Minnie.

Change-Id: I3befdefb0b70c9ffab848508f837e3db4a97b70d
Reviewed-on: https://chromium-review.googlesource.com/354351
Commit-Ready: Chinyue Chen <chinyue@chromium.org>
Tested-by: Chinyue Chen <chinyue@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>