WIP: CHROMIUM: extcon: cros_ec: Report USB Type-C cable states
Add support for and report EXTCON_TYPEC_POLARITY, EXTCON_TYPEC_PIN_ASSIGN,
and EXTCON_DISP_DP cable states if supported by the EC.

Also, always report the initial cable state even if nothing changed, and
store the vconn state for later consumption when updating the cable state.

Note: This patch is not expected to be the final version. Cable
properties (polarity, pin assignment) are not supposed to be reported
as cable states, but as separate properties. The extcon subsystem
currently does not support exchanging cable properties. Once property
support is added to extcon, the code needs to be updated to support it.

BUG=chrome-os-partner:52872
TEST=Check DP display

Change-Id: I33dd62ff389a3eb22149f60550b42496e8b40c89
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/362355
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>